Changes to implement DataBase, new repository classes, and changes in entity relations

+++ web_ejem5/src/main/java/es/codeurjc/web/controller/CastController.java
@@ -3,6 +3,7 @@ package es.codeurjc.web.controller;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.text.SimpleDateFormat;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -25,7 +26,7 @@ import es.codeurjc.web.entities.*;
 @Controller
 public class CastController {
 
-    private static final String CAST_IMAGES_FOLDER = "cast_images";
+	private static final String CAST_IMAGES_FOLDER = "cast_images";
 
     @Autowired
     private CastService castService;
@@ -42,46 +43,52 @@ public class CastController {
         return "castList_template";
     }
 
-    @GetMapping("/cast/{id}")
-    public String showCast(Model model, @PathVariable long id) {
-        Cast cast = castService.findById(id);
-        if (cast == null) {
-            return "redirect:/error?status=404&resource=cast";
-        }
-        model.addAttribute("cast", cast);
-        return "cast_template";
-    }
+	@GetMapping("/cast/{id}")
+	public String showCast(Model model, @PathVariable long id) {
+
+		Optional<Cast> op = castService.findById(id);
+		if (op.isPresent()) {
+			Cast cast = op.get();
+			model.addAttribute("cast", cast);
+			return "cast_template";
+		} else {
+			return "cast_not_found_template";
+		}
+	}
 
     @GetMapping("/cast/{id}/image")	
     public ResponseEntity<Object> downloadCastImage(@PathVariable int id) throws MalformedURLException {
         return imageService.createResponseFromImage(CAST_IMAGES_FOLDER, id);		
     }
 
-    @GetMapping("/cast/new")
-    public String newCastForm(Model model) {
-        model.addAttribute("movies",moviesService.findAll());
-        return "new_cast_template";
-    }
-    
-    @PostMapping("/cast/new")
-    public String newCast(Model model, @RequestParam(value = "castMovies", required = false) List<Long> castMovies,@RequestParam String castName, @RequestParam String castBiography,@RequestParam  @DateTimeFormat(pattern = "yyyy-MM-dd") Date castBirthDate, @RequestParam String castOriginCountry, MultipartFile castImage) throws IOException {
-        SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
-        String castBirthDateCorrect = sdf.format(castBirthDate);
-        Cast cast;
-        if (castMovies!=null){
-            List<Movie> moviesList=new ArrayList<Movie>();
-            for (int i=0;i<castMovies.size();i++){
-                moviesList.add(moviesService.findById(castMovies.get(i)));
-            }
-            cast=new Cast(castName,castBiography,castBirthDateCorrect,castOriginCountry,moviesList);
-        } else{
-            cast=new Cast(castName,castBiography,castBirthDateCorrect,castOriginCountry,null);
-        }
-        castService.save(cast);
-        
-        imageService.saveImage(CAST_IMAGES_FOLDER, cast.getId(), castImage);
-        return "cast_created_template";
-    }
+	@GetMapping("/cast/new")
+	public String newCastForm(Model model) {
+		model.addAttribute("movies", moviesService.findAll());
+		return "new_cast_template";
+	}
+
+	@PostMapping("/cast/new")
+	public String newCast(Model model, @RequestParam(value = "castMovies", required = false) List<Long> castMovies,
+			@RequestParam String castName, @RequestParam String castBiography,
+			@RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") Date castBirthDate,
+			@RequestParam String castOriginCountry, MultipartFile castImage) throws IOException {
+		SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
+		String castBirthDateCorrect = sdf.format(castBirthDate);
+		Cast cast;
+		if (castMovies != null) {
+			List<Movie> moviesList = new ArrayList<Movie>();
+			for (int i = 0; i < castMovies.size(); i++) {
+				moviesList.add(moviesService.findById(castMovies.get(i)));
+			}
+			cast = new Cast(castName, castBiography, castBirthDateCorrect, castOriginCountry, moviesList);
+		} else {
+			cast = new Cast(castName, castBiography, castBirthDateCorrect, castOriginCountry, null);
+		}
+		castService.save(cast);
+
+		imageService.saveImage(CAST_IMAGES_FOLDER, cast.getId(), castImage);
+		return "cast_created_template";
+	}
 
     @PostMapping("/cast/{id}/delete")
     public String deleteCast(Model model, @PathVariable long id) throws IOException {
@@ -90,44 +97,45 @@ public class CastController {
         return "cast_deleted_template";
     }
 
-    @GetMapping("/cast/{id}/modify")
-    public String modifyCastForm(Model model,@PathVariable long id) {
-        Cast cast = castService.findById(id);
-        if (cast == null) {
-            return "redirect:/error?status=404";
-        }
-        model.addAttribute("cast",cast);
-        model.addAttribute("allMovies", moviesService.findAll());
-        return "modify_cast_template";
-    }
-    
-    @PostMapping("/cast/{id}/modify")
-    public String modifyCast(Model model,@RequestParam(value = "castMovies", required = false) List<Long> castMovies,@PathVariable long id,@RequestParam String castName, @RequestParam String castBiography,@RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") Date castBirthDate, @RequestParam String castOriginCountry, MultipartFile castImage) throws IOException {
-        SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
-        String castBirthDateCorrect = sdf.format(castBirthDate);
-        Cast cast = castService.findById(id);
-        if (cast == null) {
-            return "redirect:/error?status=404";
-        }
-        if (castMovies!=null){
-            List<Movie> moviesList=new ArrayList<Movie>();
-            for (int i=0;i<castMovies.size();i++){
-                moviesList.add(moviesService.findById(castMovies.get(i)));
-            }
-            cast.setMovies(moviesList);
-        } else{
-            cast.setMovies(null);
-        }
-        cast.setBiography(castBiography);
-        cast.setBirthDate(castBirthDateCorrect);
-        cast.setName(castName);
-        cast.setOriginCountry(castOriginCountry);
-        
-        imageService.deleteImage(CAST_IMAGES_FOLDER, cast.getId());
-        imageService.saveImage(CAST_IMAGES_FOLDER, cast.getId(), castImage);
+	@GetMapping("/cast/{id}/modify")
+	public String modifyCastForm(Model model, @PathVariable long id) {
+		Optional<Cast> op = castService.findById(id);
+		if (op.isPresent()) {
+			Cast cast = op.get();
+			model.addAttribute("cast", cast);
+			model.addAttribute("allMovies", moviesService.findAll());
+			return "modify_cast_template";
+		} else {
+			return "cast_not_found_template";
+		}
+	}
 
-        model.addAttribute("cast",cast);
+	@PostMapping("/cast/{id}/modify")
+	public String modifyCast(Model model, @RequestParam(value = "castMovies", required = false) List<Long> castMovies,
+			@PathVariable long id, @RequestParam String castName, @RequestParam String castBiography,
+			@RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") Date castBirthDate,
+			@RequestParam String castOriginCountry, MultipartFile castImage) throws IOException {
+		SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
+		String castBirthDateCorrect = sdf.format(castBirthDate);
+		Optional<Cast> op = castService.findById(id);
+		if (op.isPresent()) {
+			List<Movie> moviesList = new ArrayList<Movie>();
+			Cast cast = op.get();
+			for (int i = 0; i < castMovies.size(); i++) {
+				moviesList.add(moviesService.findById(castMovies.get(i)));
+			}
+			cast.setMovies(moviesList);
+			cast.setBiography(castBiography);
+			cast.setBirthDate(castBirthDateCorrect);
+			cast.setName(castName);
+			cast.setOriginCountry(castOriginCountry);
+			imageService.deleteImage(CAST_IMAGES_FOLDER, cast.getId());
+			imageService.saveImage(CAST_IMAGES_FOLDER, cast.getId(), castImage);
+			model.addAttribute("cast", cast);
 
-        return "cast_modified_template";
-    }
+			return "cast_modified_template";
+		} else {
+			return "cast_not_found_template";
+		}
+	}
 }
